refactor: merged register
