+++ internal/handler/user_handler.go
@@ -185,3 +185,49 @@ func (h *UserHandler) ResetPasswordCreate(c *fiber.Ctx) error {
 
 	return c.SendStatus(fiber.StatusNoContent)
 }
+
+// Register godoc
+// @Summary Register new user
+// @Description Register new user
+// @Tags auth
+// @Accept json
+// @Produce json
+// @Param user body dto.RegisterRequestBody true "user information"
+// @Success 201 {object} dto.SuccessResponse[dto.TokenWithUserInformationResponseBody] "user successfully registered"
+// @Failure 400 {object} dto.ErrorResponse "your request is invalid"
+// @Failure 500 {object} dto.ErrorResponse "system cannot register user"
+// @Router /auth/register [post]
+func (h *UserHandler) Register(c *fiber.Ctx) error {
+	user := new(dto.RegisterRequestBody)
+	err := c.BodyParser(&user)
+	if err != nil {
+		return apperror.BadRequestError(err, "your request is invalid")
+	}
+
+	validate := validator.New()
+
+	if err := validate.Struct(user); err != nil {
+		return apperror.BadRequestError(err, "your request body is incorrect")
+	}
+	gormUser := &domain.User{
+		Email:    user.Email,
+		Username: user.UserName,
+		Password: user.Password,
+	}
+
+	accessToken, refreshToken, err := h.userService.Create(c.Context(), gormUser)
+	if err != nil {
+		return err
+	}
+
+	data := dto.TokenWithUserInformationResponseBody{
+		AccessToken:     accessToken,
+		RefreshToken:    refreshToken,
+		UserInformation: *gormUser,
+	}
+
+	res := dto.Success(data)
+
+	return c.Status(fiber.StatusCreated).JSON(res)
+
+}
